feat: Remove the Rez job environment from the generated job bundles
After this change, the queue that this submits jobs to will need either:

1. The dependent software is installed and in the system PATH for the
   workers.
2. The queue has a queue environment that accepts the RezPackages
   parameter and creates a rez environment for the render to use.

Also, put in logic to exclude the deadline_cloud_for_maya rez package
when the developer option include_adaptor_wheels is enabled.

Ran into an error loading sticky settings when removing the rez packages
from sticky settings, so rewrote the load & save logic for them to be
more simple and direct instead of using a special persistent data class
object, and to ignore problems in the sticky settings file instead of
failing to load.

Had to reduce the coverage target from 50 to 45.

Signed-off-by: Mark Wiebe <markw@amazon.com>