fix: license error string for Arnold
What was the problem/requirement? (What/Why)
A customer reported an issue where their Maya job succeeds even when `abort_on_license_fail` option is enabled and license checkout for Arnold fails. 
This is happening because the string which we are trying to match in adaptor.py line#217 is different from what is actually printing in the log. 
What was the solution? (How)
I changed the string from `aborting render because the abort_on_license_fail option was enabled` to `aborting render because this is a batch render and abort_on_license_fail option is enabled`

What is the impact of this change?
It will catch Arnold license error an fail the job. 

How was this change tested?
It is not tested. Please help me test it. I think it is easy enough to fix the string without testing. I'm not sure if it is possible. This is my first PR.

Did you run the "Job Bundle Output Tests"? If not, why not? If so, paste the test results here.
No

Was this change documented?
No

Is this a breaking change?
No

Here's the snippet of log customer shared:
```
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDOUT: 00:34:55  5452MB ERROR   |  aborting render because this is a batch render and abort_on_license_fail option is enabled
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDERR: Error: Cannot find procedure "xgmPreview".
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDERR: Error: An execution error occured in the expression xgmRefreshPreview.
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDERR: Error: Cannot find procedure "xgmPreview".
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDERR: Error: An execution error occured in the expression xgmRefreshPreview1.
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDERR: Error: Cannot find procedure "xgmPreview".
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDERR: Error: An execution error occured in the expression xgmRefreshPreview2.
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDERR: Error: ERROR   |  aborting render because this is a batch render and abort_on_license_fail option is enabled
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDOUT: MayaClient: Finished Rendering Frame 1226
2024/11/29 11:10:09-08:00 ADAPTOR_OUTPUT: STDOUT: 
2024/11/29 11:10:10-08:00 INFO: Done MayaAdaptor main
2024/11/29 11:10:10-08:00 Process pid 4380 exited with code: 0 (unsigned) / 0x0 (hex)
2024/11/29 11:10:10-08:00 ----------------------------------------------
2024/11/29 11:10:10-08:00 Uploading output files to Job Attachments
2024/11/29 11:10:10-08:00 ----------------------------------------------
```